% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aggregate_sims.R
\name{aggregate_sims}
\alias{aggregate_sims}
\title{Aggregate simulations}
\usage{
aggregate_sims(file.prefix, profiles, save.file = NULL, which.reps = 1:60,
  tol = 0.005, level = 0.1, run.cv = TRUE)
}
\arguments{
\item{file.prefix}{File prefix}

\item{profiles}{Mean values}

\item{save.file}{Optional output file}

\item{tol}{Threshold for declaring sites separated}

\item{level}{Level at which to calcluate average true and false positive rate}

\item{run.cv}{For JADE, calculate cv error if it hasn't been run already.}

\item{which.rep}{Which simulations (length B)}
}
\value{
A list with elements
#' \describe{
 \item{\code{names}}{Name of each method in order}
 \item{\code{all.stats}}{B x n x F array of statistic values for alternative methods.}
  \item{\code{site.labels}}{Which sites are separated}
  \item{\code{tol, level}}{Parameters passed in}
  \item{\code{avg.tpr, avg.fpr}}{Average true and false positive rates at given level and for JADE (14th column)}
}
}
\description{
Aggregate results of simulations
}

