% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/alternatives_binomial.R
\name{alternatives_binomial}
\alias{alternatives_binomial}
\title{Run alternative methods for binomial simulations}
\usage{
alternatives_binomial(file.prefix, bs.factor = 5, tmpdir = ".")
}
\arguments{
\item{file.prefix}{File prefixe}

\item{bs.factor}{Separation between sites used for BSmooth}

\item{tmpdir}{Directory to use for temporary files.}
}
\value{
Nothing
}
\description{
Run alternative methods BSmooth,
MethylKit and some t-test based methods on binomial data
}

