% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/alternatives_normal.R
\name{alternatives_normal}
\alias{alternatives_normal}
\title{Run alternative methods for normal simulations}
\usage{
alternatives_normal(file.prefix)
}
\arguments{
\item{file.prefix}{File prefixe}
}
\value{
Nothing
}
\description{
Run alternative t-test based methods on normal data
}

