% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/avg_by_ct.R
\name{avg_by_ct}
\alias{avg_by_ct}
\title{Get average tpr and fpr rates over replicates. Align rates for fits with the same proportion detected.}
\usage{
avg_by_ct(tpr.list, fct.list)
}
\arguments{
\item{tpr.list}{A list of vectors giving true positive rates}

\item{fct.list}{A list of vectors giving counts of false positive}
}
\value{
list with fpr, tpr and s.e
}
\description{
Get average tpr and fpr rates over replicates. Align rates for fits with the same proportion detected.
}

