% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/avg_by_prop.R
\name{avg_by_prop}
\alias{avg_by_prop}
\title{Get average tpr and fpr rates over replicates. Align rates for fits with the same proportion detected.}
\usage{
avg_by_prop(tpr.list, fpr.list, prop.list, npoints = 200)
}
\arguments{
\item{tpr.list}{A list of vectors giving true positive rates}

\item{fpr.list}{A list of vectors giving false positive rates}

\item{prop.list}{A list of vectors giving proportion discovered}

\item{npoints}{How many points to evaluate at}
}
\value{
list with fpr, tpr and s.e
}
\description{
Get average tpr and fpr rates over replicates. Align rates for fits with the same proportion detected.
}

