% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_data.R
\name{binomial_data}
\alias{binomial_data}
\title{Simulate binomial data from given read counts}
\usage{
binomial_data(profile, reads, reff.sd)
}
\arguments{
\item{profile}{Mean values (length p)}

\item{reads}{Reads per site.}

\item{reff.sd}{Standard deviation of random effects}
}
\value{
A data frame with columns y and reads
}
\description{
Simulate binomial data conditional on number of reads observed.
}

