% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_data_wrappers.R
\name{generate_data_ar}
\alias{generate_data_ar}
\title{Wrapper to generate normal auto-regressive data}
\usage{
generate_data_ar()
}
\description{
Generate normal AR  data. Run form a directory with a "data/" subdirectory.
}

