% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_data_wrappers.R
\name{generate_data_binomial}
\alias{generate_data_binomial}
\title{Wrapper to generate binomial data}
\usage{
generate_data_binomial(seed = 1111111, start.rep = 1, stop.rep = 60)
}
\description{
Generate binomial data. Run form a directory with a "data/" subdirectory.
}

