% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_data_wrappers.R
\name{generate_data_long}
\alias{generate_data_long}
\title{Wrapper to generate normal auto-regressive data with long profiles}
\usage{
generate_data_long(arsd, rho, n.rep, sample.size = c(20, 20))
}
\description{
Generate normal AR  data. Run form a directory with a "data/" subdirectory.
}

