% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_data_wrappers.R
\name{generate_data_meth}
\alias{generate_data_meth}
\title{Wrapper to generate binomial data based on methylation}
\usage{
generate_data_meth(prefix, re, sample.size, n.rep, seed = NULL)
}
\description{
Generate binomial data.
Run form a directory with a "data/" subdirectory.
}

