% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_data_wrappers.R
\name{generate_data_re}
\alias{generate_data_re}
\title{Wrapper to generate normal data with random effects}
\usage{
generate_data_re()
}
\description{
Generate normal data with random effects. Run form a directory with a "data/" subdirectory.
}

