% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_data_wrappers.R
\name{generate_data_short}
\alias{generate_data_short}
\title{Wrapper to generate normal data with short profiles}
\usage{
generate_data_short(sigma, rho, re, n.rep, prefix, sample.size)
}
\description{
Generate normal data. Run form a directory with a "data/" subdirectory.
}

