% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rates_by_region.R
\name{get_region_rates}
\alias{get_region_rates}
\title{Get average region level tpr and fpr from and agg object}
\usage{
get_region_rates(agg.obj, stat.names, merge.margin = 0, min.length = 1,
  min.acc = 0)
}
\arguments{
\item{agg.obj}{An object produced by aggregate_sims}

\item{stat.names}{Which statistics to compare to}

\item{merge.margin}{Merge regions separated by merge.margin}

\item{min.length}{Only consider regions longer than min.length}

\item{min.acc}{Not currently used}
}
\value{
A list of objects produced by get_tpr_fpr
}
\description{
Get average region level tpr and fpr from and agg object
}

