% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_data.R
\name{normal_data}
\alias{normal_data}
\title{Simulate normal data}
\usage{
normal_data(profile, sd, rho, reff.sd)
}
\arguments{
\item{profile}{Mean values (length p)}

\item{sd}{Standard deviation}

\item{rho}{Auto-regression parameter}

\item{reff.sd}{Standard deviation of random effects}
}
\value{
A length p vector of simulated data points.
}
\description{
Simulate normal data
with auto-regressive errors or random effects
}

