% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_rates2.R
\name{plot_rates2}
\alias{plot_rates2}
\title{Plot ROC curves for simulations}
\usage{
plot_rates2(rate.list, cols, ltys, main = "", lwd = 1.5)
}
\arguments{
\item{rate.list}{An object produced by}

\item{cols}{Colors}

\item{ltys}{Line types}

\item{main}{Plot title}

\item{lwd}{Line Width}
}
\value{
Nothing
}
\description{
Plot ROC curves for JADE and any other methods
}

