% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/run_jade_binomial.R
\name{run_jade_binomial}
\alias{run_jade_binomial}
\title{Run JADE for one binomial simulation}
\usage{
run_jade_binomial(file.prefix)
}
\arguments{
\item{file.prefix}{File prefix}
}
\value{
A data frame with columns y and reads
}
\description{
Run one binomial simulation with existing data
}

