% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/run_jade_normal.R
\name{run_jade_normal}
\alias{run_jade_normal}
\title{Run JADE for one normal simulation}
\usage{
run_jade_normal(file.prefix, run.f0 = TRUE, folds = 1:5,
  log.gamma.min = -5)
}
\arguments{
\item{file.prefix}{File prefix}

\item{run.f0}{Run at gamma=0}

\item{folds}{Which folds to calculate path for (0 is full data)}
}
\value{
Nothing
}
\description{
Run one normal simulation with existing data
}

