% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_data.R
\name{tile_binomial_data}
\alias{tile_binomial_data}
\title{Simulate binomial data}
\usage{
tile_binomial_data(profile, mean.reads, reff.sd, mean.tile.length = 30)
}
\arguments{
\item{profile}{Mean values (length p)}

\item{mean.reads}{Desired average number of reads per site.}

\item{reff.sd}{Standard deviation of random effects}
}
\value{
A data frame with columns y and reads
}
\description{
Simulate binomial data
by tiling reads.
}

